+++ docs/changelog.md
@@ -7,6 +7,7 @@
 - Add a sound for NAVI being aborted due to interdiction.
 - Play targetDestroyed sound when the computer reports another ship has been destroyed.
 - Add option to gag holographic avatar names in computer announcements (conf gag holo). This is disabled by default, so custom holographic avatar names will now be visible in computer announcements.
+- Add self-destruct countdown sequence sounds (60s, 30s, 10s, 5-4-3-2-1, goodbye).
 
 ### Changed
 - Wake on Red now also triggers on blue, purple, and orange alerts.